+++ WeatherStation/src/main/java/instruments/Barometer.java
@@ -5,8 +5,8 @@ public class Barometer {
      * Returns the current pressure in Millibars.
      * NOTE: The baseline atmospheric pressure is 1,013 millibars at sea level.
      */
-    public double getAtmosphericPressure() {
+    public int getAtmosphericPressure() {
         //TODO implement
-        return 0.0;
+        return 0;
     }
 }
