+++ WeatherStation/src/main/java/instruments/Hygrometer.java
@@ -3,7 +3,7 @@ package instruments;
 
 public class Hygrometer {
 
-    public double getCurrentHumidity() {
-        return 0.0;
+    public int getCurrentHumidity() {
+        return 0;
     }
 }
