+++ WeatherStation/src/main/java/instruments/Thermometer.java
@@ -2,7 +2,7 @@ package instruments;
 
 public class Thermometer {
 
-    public double getCurrentTemperature() {
-        return 0.0;
+    public int getCurrentTemperature() {
+        return 0;
     }
 }
